+++ taichi/program/program.cpp
@@ -33,8 +33,8 @@
 #include "taichi/backends/vulkan/vulkan_loader.h"
 #endif
 #ifdef TI_WITH_DX11
-#include "taichi/backends/dx/dx_program.h"
-#include "taichi/backends/dx/dx_api.h"
+#include "taichi/runtime/program_impls/dx/dx_program.h"
+#include "taichi/runtime/dx/dx_api.h"
 #endif
 
 #if defined(TI_ARCH_x64)
