+++ taichi/python/export_misc.cpp
@@ -5,7 +5,7 @@
 
 #include "taichi/runtime/metal/api.h"
 #include "taichi/runtime/gfx/runtime.h"
-#include "taichi/backends/dx/dx_api.h"
+#include "taichi/runtime/dx/dx_api.h"
 #include "taichi/common/core.h"
 #include "taichi/common/interface.h"
 #include "taichi/common/task.h"
