+++ tests/cpp/backends/dx11_device_test.cpp
@@ -5,7 +5,7 @@
 #include "taichi/ir/ir_builder.h"
 #include "taichi/backends/dx/dx_device.h"
 #include "taichi/backends/dx/dx_info_queue.h"
-#include "taichi/backends/dx/dx_program.h"
+#include "taichi/runtime/program_impls/dx/dx_program.h"
 #include "taichi/system/memory_pool.h"
 #include "tests/cpp/program/test_program.h"
